Поправил вывод количества знаков после запятой

+++ LR2/data.hpp
@@ -10,4 +10,4 @@ inline double df(double x) {
 	return(5*pow(x,4)-5.3*4*pow(x,3)+9.8*3*pow(x,2)-10*2*x+8.8);
 }
 
-const double Eps = 1E-8;
+const float Eps = 1E-8;
